Fix ImageShape.kt

+++ image-krop/src/main/kotlin/io/bashpsk/emptylibs/imagekrop/crop/ImageKropUI.kt
@@ -652,7 +652,7 @@ private fun KropShapeCustomization(state: ImageKropState) {
                     steps = 37,
                     onValueChange = { newValue ->
 
-                        val newShape = shape.copy(edges = newValue.toInt().toShort())
+                        val newShape = shape.copy(edges = newValue.toInt())
 
                         state.apply {
